requirements.txt
Django
numpy==1.21.6
opencv-python==4.7.0.72
Pillow==9.5.0
